Fixed style name always being "Style" (#117)
Styles used to be derived from the `Style` class and their class name was used in verbose output. Styles are now dynamic objects and there is no inheritance so this lead to all styles being called "Style".

This changes the way Styles are added to name them as per the configuration file or to give them a sensible default name when initialised through another means.