Adapt configuration option and improve testing
Reviewer has suggested changing the argument handling to prevent the
user from having to specify an option because this is contrary to
spirit of making the flag voluntary.  Instead, the code checks a
variety of locations before raising an exception if the configuration
cannot be found.

This also improves temporary directory handling in tests as
recommended by the reviewer, replacing clunky custom code with
a function provide pytest.